Refactor Flask app to add user authentication and registration functionality

+++ app.py
@@ -10,7 +10,20 @@ from werkzeug.urls import url_parse
 from datetime import datetime
 import os
 
+from flask_mail import Mail, Message
+
 app = Flask(__name__)
+
+from flask_mail import Mail, Message
+from flask import render_template
+
+mail = Mail(app)
+
+def send_email(subject, sender, recipients, text_body, html_body):
+    msg = Message(subject, sender=sender, recipients=recipients)
+    msg.body = text_body
+    msg.html = html_body
+    mail.send(msg)
 app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY') or 'you-will-never-guess'
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL') or 'sqlite:///app.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
@@ -53,6 +66,7 @@ class RegistrationForm(FlaskForm):
     email = StringField('Email', validators=[DataRequired(), Email()])
     password = PasswordField('Password', validators=[DataRequired(), Length(min=8)])
     confirm_password = PasswordField('Confirm Password', validators=[DataRequired(), EqualTo('password')])
+    terms_agree = BooleanField('I agree to the Terms and Conditions', validators=[DataRequired()])
     submit = SubmitField('Sign Up')
 
     def validate_username(self, username):
@@ -64,6 +78,32 @@ class RegistrationForm(FlaskForm):
         user = User.query.filter_by(email=email.data).first()
         if user:
             raise ValidationError('That email is taken. Please choose a different one.')
+
+    def validate_password(self, password):
+        if not any(char.isdigit() for char in password.data):
+            raise ValidationError('Password must contain at least one number.')
+        if not any(char.isupper() for char in password.data):
+            raise ValidationError('Password must contain at least one uppercase letter.')
+
+class ChangePasswordForm(FlaskForm):
+    old_password = PasswordField('Current Password', validators=[DataRequired()])
+    new_password = PasswordField('New Password', validators=[DataRequired(), Length(min=8)])
+    confirm_new_password = PasswordField('Confirm New Password', validators=[DataRequired(), EqualTo('new_password')])
+    submit = SubmitField('Change Password')
+
+class DeleteAccountForm(FlaskForm):
+    password = PasswordField('Password', validators=[DataRequired()])
+    confirm_delete = BooleanField('I understand this action is irreversible', validators=[DataRequired()])
+    submit = SubmitField('Delete Account')
+
+class ResetPasswordRequestForm(FlaskForm):
+    email = StringField('Email', validators=[DataRequired(), Email()])
+    submit = SubmitField('Request Password Reset')
+
+class ResetPasswordForm(FlaskForm):
+    password = PasswordField('New Password', validators=[DataRequired(), Length(min=8)])
+    confirm_password = PasswordField('Confirm New Password', validators=[DataRequired(), EqualTo('password')])
+    submit = SubmitField('Reset Password')
 
 @app.route('/')
 @app.route('/index')
@@ -128,18 +168,52 @@ def change_password():
             flash('Your password has been updated!', 'success')
             return redirect(url_for('profile'))
         else:
-            flash('Invalid old password.', 'danger')
+            flash('Invalid current password.', 'danger')
     return render_template('change_password.html', title='Change Password', form=form)
 
-@app.route('/delete_account', methods=['POST'])
+@app.route('/delete_account', methods=['GET', 'POST'])
 @login_required
 def delete_account():
-    user = current_user
-    logout_user()
-    db.session.delete(user)
-    db.session.commit()
-    flash('Your account has been deleted.', 'info')
-    return redirect(url_for('index'))
+    form = DeleteAccountForm()
+    if form.validate_on_submit():
+        if current_user.check_password(form.password.data):
+            user = current_user
+            logout_user()
+            db.session.delete(user)
+            db.session.commit()
+            flash('Your account has been deleted.', 'info')
+            return redirect(url_for('index'))
+        else:
+            flash('Invalid password.', 'danger')
+    return render_template('delete_account.html', title='Delete Account', form=form)
+
+@app.route('/reset_password_request', methods=['GET', 'POST'])
+def reset_password_request():
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+    form = ResetPasswordRequestForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data).first()
+        if user:
+            send_password_reset_email(user)
+        flash('Check your email for the instructions to reset your password', 'info')
+        return redirect(url_for('login'))
+    return render_template('reset_password_request.html', title='Reset Password', form=form)
+
+@app.route('/reset_password/<token>', methods=['GET', 'POST'])
+def reset_password(token):
+    if current_user.is_authenticated:
+        return redirect(url_for('index'))
+    user = User.verify_reset_password_token(token)
+    if not user:
+        return redirect(url_for('index'))
+    form = ResetPasswordForm()
+    if form.validate_on_submit():
+        user.set_password(form.password.data)
+        db.session.commit()
+        flash('Your password has been reset.', 'success')
+        return redirect(url_for('login'))
+    return render_template('reset_password.html', form=form)
 
 @app.errorhandler(404)
 def not_found_error(error):
@@ -149,6 +223,26 @@ def not_found_error(error):
 def internal_error(error):
     db.session.rollback()
     return render_template('500.html'), 500
+
+@app.errorhandler(403)
+def forbidden_error(_):
+    return render_template('403.html'), 403
+
+def send_password_reset_email(user):
+    token = user.get_reset_password_token()
+    subject = 'Reset Your Password'
+    send_password_reset_email_async.delay(subject, user.id, token)
+
+def send_password_reset_email_async(subject, user_id, token):
+    user = User.query.get(user_id)
+    if user:
+        send_email(subject,
+               sender=app.config['ADMINS'][0],
+               recipients=[user.email],
+               text_body=render_template('email/reset_password.txt',
+                                         user=user, token=token),
+               html_body=render_template('email/reset_password.html',
+                                         user=user, token=token))
 
 if __name__ == '__main__':
     app.run(debug=True)
